Refs #3219 WIP putting everything together. Test combinations.

+++ src/cpp/qos/ParameterList.cpp
@@ -693,7 +693,7 @@ bool ParameterList::readParameterListfromCDRMsg(CDRMessage_t& msg, std::function
                 }
                 case PID_TYPE_CONSISTENCY_ENFORCEMENT:
                 {
-                    uint16_t uKind(0);
+                    uint32_t uKind(0);
                     octet temp(0);
                     TypeConsistencyEnforcementQosPolicy p;
                     p.m_ignore_sequence_bounds = false;
@@ -705,7 +705,7 @@ bool ParameterList::readParameterListfromCDRMsg(CDRMessage_t& msg, std::function
                     valid &= plength >= 2;
                     if (valid)
                     {
-                        valid &= CDRMessage::readUInt16(&msg, &uKind);
+                        valid &= CDRMessage::readUInt32(&msg, &uKind);
                         p.m_kind = static_cast<TypeConsistencyKind>(uKind);
                     }
                     if (valid && plength >= 3)
